+++ Ex024.java
@@ -5,15 +5,15 @@ public class Ex024 {
         Scanner in = new Scanner(System.in);
 
         int n;
-        double preco;
+        float preco;
 
         System.out.print("Qual foi o numero de maças compradas? ");
         n = in.nextInt();
 
         if (n<12) {
-            preco = 1.50;
+            preco = 1.50f;
         } else {
-            preco = 1.30;
+            preco = 1.30f;
         }
         preco = preco * n;
         System.out.printf("O custo total foi R$ %.2f reais.%n", preco);
